requirements.txt
streamlit
langchain
langchain-huggingface
langchain-community
fastembed
chroma
transformers
torch
python-dotenv